requirements.txt
fastapi
uvicorn
python-multipart
c2pa-python
solana
solders
requests
jinja2
python-dotenv
Pillow
cryptography
# Database & Auth
sqlalchemy
passlib[bcrypt]
pyjwt